refactor: :hammer: removed unused elements and refactor components
