+++ app/src/main/java/com/rodrigo/deeplarva/modules/image/IPhotoHandler.kt
@@ -1,4 +1,4 @@
-package com.rodrigo.deeplarva.utils.image
+package com.rodrigo.deeplarva.modules.image
 
 import android.app.Activity
 import android.content.Intent
