--- app/src/main/java/com/rodrigo/deeplarva/ui/decorator/SpaceItemDecoration.kt
@@ -1,21 +1,0 @@
-package com.rodrigo.deeplarva.ui.decorator
-
-import android.graphics.Rect
-import android.view.View
-import androidx.recyclerview.widget.RecyclerView
-import androidx.recyclerview.widget.RecyclerView.ItemDecoration
-
-
-class SpaceItemDecoration(private val space: Int) : ItemDecoration() {
-    override fun getItemOffsets(
-        outRect: Rect,
-        view: View,
-        parent: RecyclerView,
-        state: RecyclerView.State
-    ) {
-        outRect.left = space
-        outRect.right = space
-        outRect.bottom = space
-        outRect.top = space
-    }
-}
