--- app/src/main/res/layout/item_view_picture.xml
@@ -1,54 +1,0 @@
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="120dp"
-    android:orientation="vertical" >
-
-    <ImageView
-        android:id="@+id/imgPicture"
-        android:scaleType="centerCrop"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
-    <RelativeLayout
-        android:id="@+id/rlShadow"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="#40000000">
-
-    </RelativeLayout>
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@drawable/border_background">
-
-    </RelativeLayout>
-
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:background="@android:color/transparent"
-        android:orientation="vertical">
-
-        <View
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            android:layout_weight="1"/>
-
-        <TextView
-            android:id="@+id/textViewTitle"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textSize="12sp"
-            android:textColor="@color/white"
-            android:padding="8dp"
-            android:textStyle="bold"
-            android:lineHeight="16dp"
-            android:hint="Conteo: XX\nTiempo: 00:00:00"
-            tools:ignore="MissingConstraints" />
-
-    </LinearLayout >
-
-</androidx.constraintlayout.widget.ConstraintLayout>
