--- app/src/main/res/menu/subsample.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-<!--    <item-->
-<!--        android:id="@+id/action_search"-->
-<!--        android:icon="@android:drawable/btn_radio"-->
-<!--        android:title="Search"-->
-<!--        android:orderInCategory="100"-->
-<!--        app:showAsAction="ifRoom" />-->
-<!--    <item-->
-<!--        android:id="@+id/action_settings"-->
-<!--        android:orderInCategory="100"-->
-<!--        android:title="@string/action_settings"-->
-<!--        app:showAsAction="never" />-->
-</menu>
